pipeline/requirements.txt
# AI
anthropic>=0.34.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# HTTP + scraping
httpx>=0.27.0
feedparser>=6.0.11
trafilatura>=1.12.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
lxml_html_clean>=0.1.0
playwright>=1.44.0

# Scheduling
apscheduler>=3.10.4

# Templating
jinja2>=3.1.4

# Config + utils
pyyaml>=6.0.1
python-dotenv>=1.0.1
python-dateutil>=2.9.0
